+++ cypress/integration/bridge_utils.js
@@ -16,7 +16,7 @@ function initMain(name, version) {
 }
 
 const SDK_NAME = "javascript_native_web";
-const SDK_VERSION = "24.4.0";
+const SDK_VERSION = "24.4.1";
 
 // tests
 describe("Bridged SDK Utilities Tests", () => {