+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 24.4.1
 - Added types for the SDK
+- Added a new method `set_id(newDeviceId)` for managing device id changes according to the device ID Type
 
 ## 24.4.0
 ! Minor breaking change ! For implementations using `salt` the browser compatability is tied to SubtleCrypto's `digest` method support